Python web scraper program
This program is a web scraper written in Python that goes into given URLs and finds any URLs within the webpage. Its stores these URLs into a list, which is converted to a set, then back to a list to ensure that only unique URLs are being visited. After it reads the original URL, it begins to read URLs from the list until it runs out of URLs.
It is important to note: Because of the nature of the internet, it is likely that you will get a very large or infinite list of URLs. Because of this, the amount of websites to be visited is limited to 500.