Fitness tracker program (Group project)
This is an app that tracks user's health by recording the amount of steps taken each day, the amount of water drank, happiness level, and the amount of calories eaten in a day. 

Please note:
This app is a work in progress and we  have not figured out a way to deploy it without the dependencies. Because of this, the app will not work on the user's system without the dependencies.